refactor(spark-core): remove SparkPlugin2 dependency; change SparkPluginStorage and Store to be class factories instead of classes (to avoid circular dependency)
