+++ package.json
@@ -191,7 +191,6 @@
     "require-dir": "^0.3.0",
     "shelljs": "~0.5.0",
     "sinon": "~1.14.1",
-    "spark-js-sdk--test-users": "~0.3.3",
     "time-grunt": "^1.3.0",
     "unist-util-visit": "^1.1.0",
     "vinyl": "^1.1.1",
@@ -204,6 +203,7 @@
   },
   "optionalDependencies": {
     "clarify": "^1.0.5",
+    "spark-js-sdk--test-users": "~0.3.3",
     "trace": "^2.0.2"
   },
   "browser": {
