+++ packages/spark-core/src/lib/storage/index.js
@@ -1,10 +1,8 @@
 export {
   persist,
-  waitForValue,
-  waitForValueAndPersistResult
+  waitForValue
 } from './decorators';
-export {default as Store} from './store';
-export {default as default} from './store';
-export {default as SparkPluginStorage} from './spark-plugin-storage';
+export {default as makeSparkStore} from './make-spark-store.js';
+export {default as makeSparkPluginStore} from './make-spark-plugin-store.js';
 export {default as MemoryStoreAdapter} from './memory-store-adapter';
 export {StorageError, NotFoundError} from './errors';
